First commit of library WLAC version2
[why]
With version 2 one will have possibility to compile only the symbols those
are necessary for particular task(s)

[how]
The is a possibility to close all symbols (from headers and from sources)
with the macros 'IGNORE_ALL_WLAC_SYMBOLS'. When all symols are closed, then
opening specific symbol (function or variable) should be done in following way
symbol_name_needed the macros should be debined that has the name, that begins
by symbol name and ends '_needed'.

[example]
For example if during compilation of WLAC one needs to export from library, then following should be done
(in order to export pthread_create)
#define pthread_needed
All these definations should be declared in the file wlac_symbols.h, that latter on can be used by WLAC
user to see which symbols are exported